+++ src/main/java/lotto/domain/LottoRankRecord.java
@@ -4,8 +4,8 @@ import java.util.Objects;
 
 public class LottoRankRecord {
 
-    private LottoRank lottoRank;
-    private int number;
+    private final LottoRank lottoRank;
+    private final int number;
 
     public LottoRankRecord(LottoRank lottoRank, int number) {
         this.lottoRank = lottoRank;
