+++ src/main/java/lotto/domain/judgementway/JudgementWay.java
@@ -1,4 +1,4 @@
-package lotto.domain.judgeway;
+package lotto.domain.judgementway;
 
 import java.util.function.BiFunction;
 import lotto.domain.WinningResult;
